Add edit jetpack mode

+++ app/src/main/java/com/example/lessonslist/domain/student/StudentItem.kt
@@ -2,7 +2,7 @@ package com.example.lessonslist.domain.student
 
 
 data class StudentItem(
-    val paymentBalance: Float,
+    val paymentBalance: Int,
     val name: String,
     val lastname: String,
     val group: String,
